// src/Util/calorieninjas/types.ts
export interface NutritionData {
    sugar_g: number
    fiber_g: number
    serving_size_g: number
    sodium_mg: number
    name: string
    potassium_mg: number
    fat_saturated_g: number
    fat_total_g: number
    calories: number
    cholesterol_mg: number
    protein_g: number
    carbohydrates_total_g: number
}